fix(metrics): use int64 histogram with explicit bucket boundaries for span age
Buckets [0,1,5,10,25,50,100,250,500,1000,2500,5000,10000,30000,60000] ms
cover sub-ms through 1-minute eviction range. Switches from float64 to
int64 since millisecond durations are whole numbers.

+++ processor/retroactivesampling/processor.go
@@ -33,7 +33,7 @@ func newProcessor(set component.TelemetrySettings, cfg *Config, next consumer.Tr
 		return nil, err
 	}
 	obs := func(d time.Duration) {
-		tb.RetroactiveSamplingBufferSpanAgeOnEviction.Record(context.Background(), float64(d.Milliseconds()))
+		tb.RetroactiveSamplingBufferSpanAgeOnEviction.Record(context.Background(), d.Milliseconds())
 	}
 	ic := cache.New(cfg.MaxInterestCacheEntries)
 	buf, err := buffer.New(cfg.BufferFile, cfg.MaxBufferBytes, obs)
